added option to decode headers only
If the 'pcm' parameter to mp3dec_decode_frame()
is set to NULL, all bitstream and audio decoding
is skipped, but the header is still decoded and
frame info is still generated.

This is useful for quickly scanning an MP3 file,
e.g. to determine the duration or generate a
VBR seek index.