Eliminated warning in Visual Studio 2019.
The Microsoft C/C++ compiler warned about possibly
writing too much memory in lines like this:

    memset(scratch.grbuf[0], 0, 576*2*sizeof(float));

Although the code is correct, it confuses the compiler
into thinking more memory is being zeroed out than
intended. scratch.grbuf[0] holds only half as many
bytes as scratch.grbuf, but the code intends to
zero out the entire array, not just the first half.

Replaced with identical code that eliminates
the warning:

    memset(scratch.grbuf, 0, sizeof(scratch.grbuf));

This is also easier to understand and verify as correct.